refactor: move address plugin to core services
