+++ src/registerPlugins.js
@@ -4,6 +4,7 @@
 /* node-app imports */
 /* core-services */
 import registerAccountsPlugin from "./core-services/account/index.js";
+import registerAddressPlugin from "./core-services/address/index.js";
 import registerCatalogPlugin from "./core-services/catalog/index.js";
 import registerCartPlugin from "./core-services/cart/index.js";
 import registerCorePlugin from "./core-services/core/index.js";
@@ -22,7 +23,6 @@ import registerTagsPlugin from "./core-services/tags/index.js";
 import registerTaxesPlugin from "./core-services/taxes/index.js";
 
 /* plugins */
-import registerAddressPlugin from "./plugins/address/index.js";
 import registerLegacyAuthorizationPlugin from "./plugins/legacy-authorization/index.js";
 import registerCheckoutPlugin from "./plugins/checkout/index.js";
 import registerDashboardPlugin from "./plugins/dashboard/index.js";
